requirements.txt
colorama
discord.py
aiosqlite
numpy